Remove org-link-types
* lisp/org.el (org-link-types): Remove this variable.

* testing/lisp/test-ox.el (test-org-export/custom-protocol-maybe):

Remove copy-sequence in the tests. Actually it seems like the let
statement is no longer needed. I guess it was around to protect the list
from extra links getting defined. A downside of this is the links may
get added during tests. That does not seem like a big deal if those are
run in batch mode.

* lisp/org.el: The `org-link-types' variable was replaced with the
  `org-link-types' function everywhere.